requirements.txt
apache-airflow-providers-google
apache-airflow-providers-docker
apache-airflow-providers-fab
duckdb==1.4.0